+++ gateway-js/package.json
@@ -26,7 +26,7 @@
   },
   "dependencies": {
     "@apollo/federation": "file:../federation-js",
-    "@apollo/query-planner-wasm": "file:../query-planner-wasm",
+    "@apollo/query-planner": "file:../query-planner-js",
     "@types/node-fetch": "2.5.4",
     "apollo-graphql": "^0.6.0",
     "apollo-reporting-protobuf": "^0.6.0",
