+++ gateway-js/src/__tests__/buildQueryPlan.test.ts
@@ -3,14 +3,14 @@ import gql from 'graphql-tag';
 import { buildQueryPlan, buildOperationContext } from '../buildQueryPlan';
 import { astSerializer, queryPlanSerializer } from 'apollo-federation-integration-testsuite';
 import { getFederatedTestingSchema } from './execution-utils';
-import { WasmPointer } from '../QueryPlan';
+import { QueryPlannerPointer } from '@apollo/query-planner';
 
 expect.addSnapshotSerializer(astSerializer);
 expect.addSnapshotSerializer(queryPlanSerializer);
 
 describe('buildQueryPlan', () => {
   let schema: GraphQLSchema;
-  let queryPlannerPointer: WasmPointer;
+  let queryPlannerPointer: QueryPlannerPointer;
 
   beforeEach(() => {
     expect(
