+++ query-planner-js/.gitignore
@@ -1,0 +1,2 @@
+# JS: Ignore the compiled output.
+dist/
