+++ query-planner-js/src/__tests__/testSetup.ts
@@ -1,0 +1,1 @@
+import 'apollo-federation-integration-testsuite';
