+++ query-planner-js/tsconfig.json
@@ -6,8 +6,5 @@
   },
   "include": ["src/**/*"],
   "exclude": ["**/__tests__", "**/__mocks__"],
-  "references": [
-    { "path": "../federation-js" },
-    { "path": "../query-planner-js" },
-  ]
+  "references": []
 }
