+++ query-planner-js/tsconfig.test.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../tsconfig.test.base",
+  "files": [
+    "./src/__tests__/testSetup.ts"
+  ],
+  "include": ["**/__tests__/**/*", "**/__mocks__/**/*"],
+  "references": [
+    { "path": "./" },
+    { "path": "../federation-integration-testsuite-js" }
+  ]
+}
